system_api: Add powerwash flag to update status
Add a powerwash flag to the update status propagated from update
engine to chrome. This will ensure that the user is informed of a
pending powerwash exactly when it is going to happen.

BUG=chromium:1070563
TEST=cros_workon_make --board=amd64-generic system_api
channel change and update on test device

Change-Id: Ic447a036ccd78142eae50c63ea80e11f666afd1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2187671
Tested-by: Miriam Polzer <mpolzer@google.com>
Commit-Queue: Miriam Polzer <mpolzer@google.com>
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Cr-Mirrored-From: https://chromium.googlesource.com/chromiumos/platform2
Cr-Mirrored-Commit: 2cfbe1dd058e6483c3b7bdd3d66748637481f5d8