shill: wifi: Remove unused "preferred device" functionality
Prior to this change, WiFiService had a PreferredDevice property that
could be set to an interface name in order for the WiFiService to
always connect using that interface when possible. Not only was this
functionality unused, but if we do have multiple wifi interfaces, we
can't rely on deterministic interface naming.

This change simply removes the property and associated
functionality. Note that since the SaveString call actually fails to
ever save the property in storage, we don't need to undergo a storage
property deprecation.

This effectively reverts CL:261210.

BUG=None
TEST=All unit tests are passing.

Change-Id: I707b6cbe329556b06c2800d5cc2adcd6bec62af2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2219291
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Cr-Mirrored-From: https://chromium.googlesource.com/chromiumos/platform2
Cr-Mirrored-Commit: 22dd19d631523f6a19a485dbc92fab485bb8ed93