vm_tools: Make concierge boot vm's that are DLCs.
This change allows concierge to query dlcservice in order to determine
the paths of the rootfs and kernel. This is intended to rework
crrev.com/c/2329207.

We maintain the behaviour that trusted vm's must not take their path
directly from the proto. We instead assert that dlcservice is trusted
equivalently to concierge itself, and so asking it for a path conveys
"trust"

BUG=chromium:953544
TEST=patch in crrev.com/i/3226524 and boot the VM

Cq-Depend: chromium:2377042
Change-Id: I82487874bc1626018f10639338e3e43ed4025456
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2367298
Tested-by: Nic Hollingum <hollingum@google.com>
Commit-Queue: Nic Hollingum <hollingum@google.com>
Reviewed-by: Abhishek Bhardwaj <abhishekbh@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Fergus Dall <sidereal@google.com>
Auto-Submit: Nic Hollingum <hollingum@google.com>
Cr-Mirrored-From: https://chromium.googlesource.com/chromiumos/platform2
Cr-Mirrored-Commit: 688989880123b36444654c0dbd9586ab5c287818