seneschal: Add support for mounting smbfs paths into Crostini
Directories within smbfs shares, or the entire share, are mounted in
the VM under the "SMB/<share-id>" directory. This allows multiple smbfs
shares, or multiple paths from a single share, to be shared into the VM
without encountering namespace conflicts.

Added proto field to SharePathRequest to capture the unique, stable
mount name for the smbfs share that the shared path belongs to. At
runtime, seneschal uses this to determine the path within the host and
also the path within the VM that the file / directory should be shared
under.

BUG=chromium:962924
TEST=Share root and/or sub-folder of share into Crostini from Files.app
TEST=Unshare shared paths via "Manage Linux Sharing"
TEST=Unmount smbfs share for a shared path

Change-Id: Ia2f040b2f69003f2297336c4f6e26f2bba1b6029
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2434711
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: Josh Simmons <simmonsjosh@google.com>
Commit-Queue: Josh Simmons <simmonsjosh@google.com>
GitOrigin-RevId: 59cccc4f9ca4a82ab8a6f3fda2652b966ed135ab