system_api: Add fields to ConnectNamespace
This patch adds traffic source and vpn routing semantic fields to
ConnectNamespaceRequest.

Note that current callers of this API will not be impacted by this
change but will need to set the |route_on_vpn| flag explicitly once the
functionality is implemented, to maintain the current behavior.

BUG=b:171363265
TEST=emerge

Change-Id: I240b228a6a3124787fd8092cc97be3a9d4b5d35a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2568912
Commit-Queue: Garrick Evans <garrick@chromium.org>
Tested-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
GitOrigin-RevId: 5c127cf7451aae14fa1534552d1522dc30f87483