arc: Add boot property for Play Auto Install generation
This will be used to enable additional options in ARC (eg, making
Web APKs appear in the launcher) when a developer is generating a new
Play Auto Install roster.

This flag is only expected to be used by Google developers on their test
devices, and so the property is only set when requested by Chrome.

BUG=b:177035416
TEST=cros_run_unit_tests --board $BOARD --packages arc-setup \
        --packages chromeos-login

Change-Id: Iee7dbb8a5680cfded213c4d54591db08ac304f6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2626548
Tested-by: Tim Sergeant <tsergeant@chromium.org>
Commit-Queue: Tim Sergeant <tsergeant@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
GitOrigin-RevId: 109363d0a03be594cf5611441597c607960c4f3d