concierge: accept repeated enum for dbus fds
The previous interface was somewhat brittle and required looking up the
documentation to determine ordering. This, on the other hand, embeds the
ordering in the code and makes it possible to deprecate passed FD types
in the future if needed.

I went ahead and just replaced the whole interface in one pass since it
seems only things internal to chromiumos call this method at this point,
and these should be upgraded in lock step (I think, please check me on
this because I'm still familiarizing myself with chromeos).

BUG=b:176120200
TEST=deployed to eve device; ensured fds are still properly passed;
ensured crostini continued to work

Cq-Depend: chrome-internal:3581217
Change-Id: Ibff813474ccc6fe84047a236490ab5d151b70f51
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2644200
Tested-by: Mike Gerow <gerow@google.com>
Commit-Queue: Mike Gerow <gerow@google.com>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Fergus Dall <sidereal@google.com>
GitOrigin-RevId: b6f9f016e8fa7fa4b806f277b15b7609259e0296