system_api: Add AttestationConveyancePreference to WebAuthn MakeCredential
During WebAuthn MakeCredential requests, the browser needs to indicate
which kind of attestation is desired. For example, for corp services,
g2f is desired.

If g2f is preferrred and u2fd is in g2f mode, then g2f attestation is
returned.
In all other cases, u2f attestation is returned. (The browser will erase
the u2f attestation if it's not wanted.)

In the API we keep an option for "None" in case we need to do it
differently in the future.

BUG=b:175158241
TEST=build system_api

Change-Id: I34fdba326c910aefc61dc6fbf70660599edd2807
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2654351
Tested-by: Yicheng Li <yichengli@chromium.org>
Commit-Queue: Yicheng Li <yichengli@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
GitOrigin-RevId: b6d713a9fb629079defc9769afd42db8b105a94f