system_api: Add method for PeripheralDataAccessEnabled
typecd will be using this method, implemented by the
Chrome features service, to determine whether a peripheral should switch
to Thunderbolt mode (when the setting is enabled), or stay in
USB+DisplayPort alternate mode (when the setting is disabled). Chrome
controls the toggling of this device setting via Chrome settings.

Add the constant so that it can become visible to Chrome (after rolling
the system_api deps).

BUG=b:182835509
TEST=None

Change-Id: I7ef8e5c851a7b1953cc04914d68ff5845c2b42b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2785966
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Prashant Malani <pmalani@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>
GitOrigin-RevId: f1d9385120e02c3f7d856ac6386b35ed96c5687c