dlp: Add D-Bus calls to check file permission
In order to implement processing of file access request
in DLP Files daemon we need to check whether access to the
file can be restricted to any DLP rule.

Following options might be:
1) The file is not known by DLP database - then it's not
downloaded and not protected by DLP.
2) The file is downloaded, but there is no DLP rule that
may protect its source.
3) The file downloaded and matched at least one blocking
DLP rule - then the file operation should be prevented
unless it will be explicitly allowed.

This requires 2 policy checks to be implemented and as it
was decided that it's too complicated to copy and implement
DLP Rules Manager in Chrome OS, for now we'll introduce a
new D-Bus API to call into Chrome's DlpRulesManager:
1) IsRestricted - whether the exact operation should be restricted;
2) IsDlpPolicyMatched - whether the file could be restricted.

It's a temporary trade-off for simplicity of the implementation,
in the future for performance improvement the policy checks can
happen in Chrome OS daemon.
This is tracked by crbug.com/1169888#c1

BUG=chromium:1169888,chromium:1169882
TEST=Build and deploy, configure DLP Files policy

Change-Id: Ib47eb437b9bf9f48440bdcc03ce56f4f9e7b4737
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2850160
Commit-Queue: Sergey Poromov <poromov@chromium.org>
Tested-by: Sergey Poromov <poromov@chromium.org>
Reviewed-by: Aya Elsayed <ayaelattar@chromium.org>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: 9232e3ff80e589709d5d7d799190d9080fb98e84