crosh: Add wireguard commands (part 1)
This patch introduces wireguard commands into crosh, which talk with
Shill via D-Bus to configure and control WireGuard VPN services.

This patch adds "list" (or "show" with no parameter) subcommand. In this
command, the program will:
1) call GetProperties() on Manager, to get the list of service paths;
2) for each service path, call GetProperties() to get the service
   properties on that path;
3) for each group of service properties, find if it represents a wg vpn
   service, if so, translate it into the internal representation.

Note that values of PrivateKey and PresharedKey of a service will not be
transmitted via D-Bus, and thus will not be displayed to the users.

BUG=b:177877310
TEST=build; manually tested with later patches

Cq-Depend: chromium:2936252
Change-Id: I3b8ef7fe3e21e0cf04a97eea3b38ee0d1a17c391
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2936742
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: db049b17579c5dd70b9490faf95e947b41e4bdcb