rmad: Initialize state handler for calibration
Before calibration, select the components to be calibrated. If any
components need to be calibrated, we should wait for the technician to
set up the equipment for calibration. At the same time, we perform a
calibration procedure (without actual calibration and monitoring)
with callbacks to update the status and progress. Once the calibration
is complete (complete or failed), the UI will receive the signal.
Then the technician can go to the next state (provide equipment or
try again).

BUG=b:187196710
TEST=emerge-zork chromeos-base/system_api rmad;
     cros_run_unit_tests --board zork --packages rmad

Change-Id: I7fd62302229ab8715f588a61ec98cc05bb90c5d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2952749
Tested-by: Gene Chang <genechang@google.com>
Auto-Submit: Gene Chang <genechang@google.com>
Reviewed-by: Cheng-Han Yang <chenghan@chromium.org>
Commit-Queue: Gene Chang <genechang@google.com>
NOKEYCHECK=True
GitOrigin-RevId: 5d58458e892417e879ebd671a02ba24956ba2cdb