power: add GetBatteryState method call
In order to be compatible with fwupd, the GetBatteryState
method call has been added to the power_manager. It is
redundant with GetPowerSupplyProperties in terms of what
information it gives about the system, but it gives the
information in the form of a tuple with three elements
instead of in the form of a protobuf, which is how
GetPowerSupplyProperties returns it. This way, fwupd does
not need to have a dependency on protobuf, and it is not
affected by potential changes to orderings or added entries
in the method call it calls.

The three properties returned by GetBatteryState are the
external power type, the battery state, and the display
battery percentage. External power type is used so that
"battery discharging" and "AC power connection" can be
viewed as separate events. The battery state is sent as
an enum matching the upower enum implementation. This
way, the method call is not dependent on system_api
changes not directly related to the method call. The
display percentage was chosen over the actual percentage
since they are very close in number, but it is easier to
test manually with the display percentage and it would be
helpful in the case that fwupd develops a UI.

BUG=b:192392680
TEST=from ssh: dbus-send --print-reply --system --type=method_call --dest=org.chromium.PowerManager /org/chromium/PowerManager org.chromium.PowerManager.GetBatteryState
  output of test:
   uint32 1
   uint32 4
   double 100
from chroot: emerge-volteer power_manager ; cros_workon_make power_manager --test --board=volteer

Change-Id: I2b16324e2f498b4f9e1ec58ca576c0abe68f92a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3003571
Reviewed-by: Twain Byrnes <binarynewts@google.com>
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Reviewed-by: Benson Leung <bleung@google.com>
Reviewed-by: Daniel Campello <campello@chromium.org>
Reviewed-by: George Popoola <gpopoola@google.com>
Tested-by: Twain Byrnes <binarynewts@google.com>
Commit-Queue: Daniel Campello <campello@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: 026895625dcd24075cded100ad26e78d61cee43a