power: Add suspend flavor to RequestSuspend
Add yet another optional parameter to the RequestSuspend message: a
suspend flavor. The default for this flavor is 0, which says "do
whatever you want powerd!" Other valid values are 1 to force a
suspend-to-RAM, and 2 to force a suspend-to-disk. Passing 2 to a system
with disable_hibernate set will downgrade back to suspend-to-RAM.

Plumb this flavor down to connect it with the low layer hibernate
capabilities. This enables a few things, including 1) Chrome to request
hibernate specifically if we wired up the UI during teamfooding, and
2) hibernate tests can now be written.

BUG=b:192360150
TEST=FEATURES=test emerge-volteer power_manager, test
powerd_dbus_suspend manually

Change-Id: I275935280eff71f28f33196afa6d1eb37e7e53c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3042620
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: 5831f6ef25487b878bda92647b81dcf28627f8c8