cryptohome: Add dbus api GetWebAuthnSecretHash
Currently, u2fd requests the WebAuthn secret from cryptohome on init,
then hashes it and keeps it in memory. The WebAuthn secret is designed
to be a secret value so cryptohome clears it after a period of time
whenever it is unlocked. However, u2fd only needs its hash on init. Make
cryptohome calculate the WebAuthn secret hash when the secret is
unlocked, store in memory, and add a dbus api for u2fd to request the
WebAuthn secret hash on init, instead of requesting the secret itself.
This can eliminate the need for a long timer in cryptohome that clears
the WebAuthn secret due to the long process of the first mount, increasing
security.

BUG=b:193199466
TEST=tast run $DUT u2fd.WebauthnUsingPIN

Change-Id: I288203b97be33eb9ce88ef502e429f217614a01d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3222982
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>
NOKEYCHECK=True
GitOrigin-RevId: 22de8900b408040eeb3245e4ff1f6540f1dd13a6