system_api: Add a wayland socket to the VirtualMachineSpec
We add the ability for users to override the wayland socket which their
VM will use when invoking concierge.

It is necessary to have this customization at the level of the proto,
since overriding concierge's environment (which is the normal way to
adjust wayland socket paths) will apply to all VMs spawned by concierge,
and we want to enable per-vm customization.

Necessary for go/secure-exo-ids.

BUG=b:200896773
TEST=n/a

Change-Id: I240eb758023b7e48d85bcd09a3d720478fb4d9bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3252859
Tested-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Commit-Queue: Nic Hollingum <hollingum@google.com>
NOKEYCHECK=True
GitOrigin-RevId: 6dfe52c30e22f27819547bc21e995176da572ab7