typecd: Send notification when USB speed is limited by the cable.
Updates the type-c daemon and dbus interface to send a notification when
a device connected to the USB-C port supports a higher USB generation
than the cable used to make the connection. This notification will only
be sent when both the cable and partner product_type_vdo1 give the USB
generation and the cable is not captive.

BUG=b:211209807
TEST=Working interactively and passing unit tests

Change-Id: I83d4a31e7bff6439355731afef3747feaea9b950
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3375490
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Reviewed-by: Benson Leung <bleung@google.com>
Commit-Queue: Jameson Thies <jthies@google.com>
Tested-by: Jameson Thies <jthies@google.com>
NOKEYCHECK=True
GitOrigin-RevId: ba74773d36bfac987207380e0ac40c5364444387