shill: wifi: enable/disable interworking selection.
As Passpoint is still an experimental feature, a flag will be used to
opt-in. When the feature is disabled, Shill accepts new set of
credentials but prevent interworking selection.  Thus it won't connect
to any compatible network using the credentials.  When it's enabled, it
performs interworking selection after scans as expected.

This CL adds a boolean property to the Wi-Fi device that allow to enable
Passpoint interworking selection. The feature is disabled by default
(boolean property set to false) and can be enabled through a D-Bus
SetProperty() call.

BUG=b:162106001
TEST=unit test;
Setup a Passpoint compatible access point and push matching credentials
to Shill using gdbus call. Ensure the device does not connect to the
access point after a scan.
Then enable the feature calling SetProperty on the WiFiDevice:
 gdbus call --system --dest org.chromium.flimflam --object-path /device/wlan0 \
    --method org.chromium.flimflam.Device.SetProperty \
    '@s "PasspointInterworkingSelectEnabled"' '@v <false>'
Check the device connects to the access point after the next scan.

Change-Id: I2fb8e2bcf84f328b97a6ac0dd21db861129da899
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3447427
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Damien Dejean <damiendejean@google.com>
Commit-Queue: Damien Dejean <damiendejean@google.com>
NOKEYCHECK=True
GitOrigin-RevId: 19cca22f5c086d015b15bb439b987d5d20e6b8ed