resourced: Introduce a new dbus message to override Memory Margins
As described in b/226425011 we hope to start experimenting with
more dynamic memory margins. This CL introduces a new DBUS
message which allows for the margins to be changed at runtime.

There will be a counter-part Chrome change which will set these
values based on a finch experiment.

BUG=b:226425011
TEST=Betty VM w/ manual dbus verification

Change-Id: Iceaf46a05570c2a4683c36062ade961f16c57a81
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3545288
Reviewed-by: Shah Hossain <shahadath@google.com>
Tested-by: Brian Geffon <bgeffon@chromium.org>
Auto-Submit: Brian Geffon <bgeffon@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: def4026bf02942bc14897e01ebb107cc978ef48f