vm_tools: Support CPU_RESTRICTION_BACKGROUND_WITH_CFS_QUOTA_ENFORCED
Previously, Concierge itself automatically controlled when to apply
cfs_quota restrictions but it turned out that Concierge was not the
best place to do so. First, cpu.shares is controlled by Chrome and
the current way is not consistent with that, and second, Concierge
cannot know all the system state e.g. Android VPN usage.

This CL is to move the feature to the Chrome side by supporting the
new enum value.

BUG=b:205762950
TEST=arc.Boot.vm, confirm the quota is still -1 (unlimited)
TEST=manually send CPU_RESTRICTION_BACKGROUND_WITH_CFS_QUOTA_ENFORCED
 from Chrome, confirm the quota is initially 25% until ARC app is
 launched.

Change-Id: Ie1d4453b7c6f8053600283fab60a204ca9722a23
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3564880
Tested-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Yusuke Sato <yusukes@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: 2b2d31be2df8986e1ada455a35fe62682c31f682