cryptohome: Add SetMediaRWDataFileProjectInheritanceFlag
Crosvm will use this DBus method to set project inheritance flag.
When enabled, the project inheritance flag (FS_PROJINHERIT_FL)
allows directory content to inherit the project id of the directory.
This project inheritance flag is usually set by vold in ARCVM through
virtio-fs.
In the older kernel, this flag can be set in user namespace. However,
in the newer kernel, this flag needs to be set in the init namespace.
Hence, we needs to delegate the flag setting from virtio-fs to
another process in the init namespace. This CL add the DBus method in
cryptohome so that the flag setting is possible from virtio-fs.

BUG=b:226908517
TEST=unit tests

Change-Id: Ic37f5805afb1fecc047f050f3981d60fd17565f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3585107
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Tested-by: Risan <risan@chromium.org>
Commit-Queue: Risan <risan@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: f35997ddcce1e55123d22c9f26a71a152096ddb0