shill: Remove --accept-hostname-from flag
This flag controls on which interfaces we should accept hostname from
DHCP. Since this is only settable via command line when starting shill,
removing this should have no real effect.

Also remove the |accepted_hostname| property in IPConfig::Properties.
This property is set in DHCPController and also exposed in D-Bus
interface, but it is not used anywhere (expect for the removed code in
this patch).

BUG=b:229349972
TEST=unit tests

Change-Id: Ic6ca28a21eb44a6a2a68cb01c594e51fbc7bcb18
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3600973
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: 0d8b2f4b4e8671d0adbf3fefed03cf2dc9c515a3