vm_tools: Support attaching USB devices to a specific container
Currently, USB devices passed into the Crostini VM are shared between
all containers. This can lead to confusing behaviour, so we implement a
method of passing a USB device to a specified container.

This CL contains changes to the Cicerone to call the implementation of
this feature in Tremplin, and changes to the vmc command line to use
this feature.

BUG=b:229913205
TEST=FEATURES=test emerge-tatl vm_protos tremplin
TEST=FEATURES=test emerge-volteer vm_protos chromeos-base/system_api
     vm_host_tools crostini_client

Cq-Depend: chromium:3744724
Change-Id: I2bdd16db18a17fbf20f143d49dae7b89dd1c4af8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3716312
Reviewed-by: David Munro <davidmunro@google.com>
Tested-by: James Ye <jamesye@google.com>
Commit-Queue: James Ye <jamesye@google.com>
Reviewed-by: Nicholas Verne <nverne@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: a7bd535adf60b111f39925f3106150200ba9c18b