cryptohome: Prepare hibernate secret in AuthSession flow
In order for hibernate to migrate properly to LVM, we need a point in
the boot flow after authentication has completed, but before the user's
homedirs are mounted. The new AuthSession flow makes this particular
execution point available to Chrome (as compared to the old auth flow in
which Chrome simply called cryptohome->Mount() with credentials).
However, at this point in the boot flow, cryptohome has not yet prepared
the hibernate secret, so hiberman fails when it attempts to retrieve it.

Add an input parameter to the GetHibernateSecret() dbus method that
allows hiberman to hand in an auth_session_id that it got from Chrome.
Cryptohome can then look up this auth session ID and compute the
hibernate secret on the fly.

The long term plan for the hibernate secret is to store it in USS. This
change allows us to split out that migration into this initial step
(moving to AuthSession), and a future step to actually use USS.

BUG=b:233103309
TEST=hibernate and resume, observe hiberman succeed

Change-Id: If4760e09a762a460d8271b0d78a5153701e5e4f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3718853
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: 8881d8100aa1575edea3c0c26cca3c771942e6c8