shill: Make portal urls writable
The captive portal tast tests need to be able to set the portal https url. This will allow a self signed cert to be installed and valid when used with an https server. The portal detector will then set state depending on the response of the https response. Since we are making https writable, we should also allow it for the primary http url. Along with this, we add a fallback variable for https urls to dbus and shill for writing and reading the properties through dbus.

BUG=b:237816536
TEST=unit tests

Change-Id: Idbc26dd88518f3df72e5703159ffec2bfa8121ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3759988
Tested-by: Michael Rygiel <michaelrygiel@google.com>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Michael Rygiel <michaelrygiel@google.com>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: 2f99b173d8272ba95bacecb5880dbc7ed13193eb