vm_tools: Make concierge clean up a dead sibling VM
This change notifies concierge when a sibling VM dies from underneath it
on the hypervisor. Prior to this it was unaware of when a sibling VM
died and its internal state wasn't updated.

BUG=b:238976461
TEST=Start a sibling VM and manually kill it on the hypervisor. Ensure
that no VVU device processes remain in the Chrome OS VM and it calls
StopVm.

Change-Id: Ie50b5631e4c477979fb2364149dfdcd5b853b806
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3931530
Auto-Submit: Abhishek Bhardwaj <abhishekbh@chromium.org>
Commit-Queue: Abhishek Bhardwaj <abhishekbh@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Tested-by: Abhishek Bhardwaj <abhishekbh@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: 67dffc38d7054dd8108fc72d34c4cd0ce66141ea