fusebox: implement Create for the client
There was some existing client code for Create, but it spoke a vanilla
D-Bus protocol with the Fusebox server (which has not yet implemented
Create support). More recently, we prefer protobufs-over-D-Bus, as it
gives us greater flexibility for protocol changes.

Read-only enforcement has also generally moved from client to server, in
the time since the existing Create client code was checked in.

This is one half of a multi-repo change. The other half is at
https://crrev.com/c/4036586

BUG=b:255520194,b:255703917
TEST=manual

Change-Id: I08c1c77ff03f68f05b90e474e5a0a8c03fdb3f0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4035867
Reviewed-by: Ben Reich <benreich@chromium.org>
Tested-by: Nigel Tao <nigeltao@chromium.org>
Commit-Queue: Nigel Tao <nigeltao@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: 8d4995452c97af07bfc791f923f8545da8cef2d8