fusebox: implement Write2 for the client
There was some existing client code for Write, but it spoke a vanilla
D-Bus protocol with the Fusebox server (which has not yet implemented
Write support). More recently, we prefer protobufs-over-D-Bus, as it
gives us greater flexibility for protocol changes.

This is one half of a multi-repo change. The other half is at
https://crrev.com/c/4037223

BUG=b:255520194,b:255703917
TEST=crrev.com/c/4037323

Change-Id: I8f8962b5b7970d6f901a2379e0b81d563b2bd860
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4035874
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Nigel Tao <nigeltao@chromium.org>
Tested-by: Nigel Tao <nigeltao@chromium.org>
Reviewed-by: Ben Reich <benreich@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: 7958b75da290dcf9a2b5e3ff9129c81eb1ff5a0e