cryptohome: Decouple data_migrator from UserDataAuth.proto
This patch removes the dependency on UserDataAuth.pb.h from
data_migrator::MigrationHelper by changing the type of
MigrationHelper::ProgressCallback. This is to avoid requiring dependency
on UserDataAuth.pb.h from the future users of MigrationHelper, which are
not necessarily in cryptohome.

BUG=b:258402655
TEST=FEATURES=test emerge-hatch cryptohome
TEST=tast run <DUT> cryptohome.VaultMigration.fscrypt_v1
TEST=tast run <DUT> cryptohome.VaultMigration.fscrypt_v2

Change-Id: Ifcb2f97acd2545760b0b678f019a801fbf5d5d9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4046082
Tested-by: Momoko Hattori <momohatt@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Commit-Queue: Momoko Hattori <momohatt@chromium.org>
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: 2ad8e9ddb21041c4920091bf0afacfeaf4edfddc