vm_tools: Add allocation_type field to CreateDiskImageRequest
Currently, Concierge's Service::CreateDiskImage() creates a sparse disk
if and only if the |disk_size| field of the request is 0 (unspecified).
The size of such a disk image is calculated as:
1. 95% of total disk space if storage ballooning is enabled
2. 90% of free disk space if storage ballooning is not enabled

ARCVM makes use of a sparse disk image when virtio-blk is used for its
/data. Since storage ballooning is not enabled for ARCVM, its disk size
is calculated based on (2).
However, the size can be too small when we perform a migration from
virtio-fs /data to virtio-blk /data. This is because when the migration
source (virtio-fs /data) is large compared to the free disk space, the
destination (disk for virtio-blk /data) can be smaller than that, in
which case the migration fails with an out-of-space error.

This CL adds a new field |allocation_type| to CreateDiskImageRequest to
allow the caller to specify whether the image should be sparse or not.
It can be combined with |disk_size|, and thus now the caller can specify
the size of a new sparse disk image.

To preserve the behaviors for existing callers, when the field is AUTO
(the default value), everything works in the way as it does today, i.e.,
a sparse disk image is created if and only if |disk_size| is 0.

When |allocation_type| is PREALLOCATE, we require |disk_size| to be
non-zero, because there is no reasonable way to calculate the default
value applicable to all use cases.

Also, currently, when there is an existing disk, Concierge resizes the
disk if |disk_size| is 0 (i.e., a sparse disk image is requested) and
|storage_ballooning| is false, unless kDiskImageUserChosenSizeXattr is
set. The xattr is set for Crostini disk images to indicate that the disk
image is created as a preallocated one with a non-zero |disk_size|.
Actually, the xattr is currently used in StartVm() as an indicator of
preallocated (non-sparse) disks.

This CL makes the xattr explicitly specific to preallocated disks, i.e.,
it is set only when 1) |allocation_type| is AUTO and |disk_size| is
non-zero (current behavior), or 2) |allocation_type| is PREALLOCATE
(which implies |disk_size| is non-zero). The variable name of the xattr
is modified accordingly.

BUG=b:232176243
BUG=b:271062919
TEST=Confirm there is no behavior change with |allocation_type| = AUTO.
TEST=Confirm that SPARSE can be combined with a non-zero |disk_size|.

Change-Id: I6f9d781d3b409d0dcf7068fb3db423ed762b365a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4297723
Reviewed-by: Abhishek Bhardwaj <abhishekbh@chromium.org>
Tested-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Satoshi Niwa <niwa@chromium.org>
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: 95c759b626ef600b6fa80e02042174b6fa1e4723