attestation: Include the ADID in enrollment certificate requests
The bruschetta project has introduced a need for knowledge of ADIDs
outside of Zero-Touch Enrollment, which we are solving by adding them
to the enrollment certificate. See go/bruschetta-adids for more
details.

This requires a client-side change to send the ADID in the
AttestationCertificateRequest (along with the SN bits quote we already
send) when requesting an enrollment cert. The server-side of this is
implemented in cl/521700419.

BUG=b:268367993
TEST=unit tests

Change-Id: I1c142cc693c996f37b404851beb4ffd27992cdb2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4387676
Commit-Queue: Fergus Dall <sidereal@google.com>
Tested-by: Fergus Dall <sidereal@google.com>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: 07e8dc179631962783b52217366d5404d3699766