system_api: Add key_profile field to SignEnterpriseChallengeRequest
Add the key profile to SignEnterpriseChallengeRequests to be
able to explicitely distinguish between different types of keys
instead of relying on the username being empty or not. This change
is required to extend the types of keys used for signing enterprise
challenges (go/remote-attestation-unmanaged-devices-dd).

BUG=b:277709127
TEST=N/A

Change-Id: Ie404e3cb33e6fe846ee6b2434730ed4db80cd4ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4414457
Tested-by: Leon Masopust <lmasopust@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Leon Masopust <lmasopust@google.com>
NOKEYCHECK=True
GitOrigin-RevId: 30e2fbff9c8ee8f191b59e552fed1ff922df95ed