system_api: Add broadcast ID to StartAuthSession reply
This adds a new ID to the session, separate from the session ID, that is
safe to use as an identifier for the session without granting the
ability to act on the session. Think of it as the being like a public
key, in contrast to the auth session ID which is the private key.

BUG=b:260904713
TEST=built protos

Change-Id: Ifb999ebb99c38213fecc32567b323d1e1ca9d737
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4448926
Reviewed-by: Behnood Momenzadeh <behnoodm@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: af58f5d37c525dcc8111a72ffdb5fd686b0fdb3d