resourced: Add VM booting mode
Add the VM booting mode which will be enabled when a VM is starting up.
If the resourced receives the 'VmStartingUpSignal' from vm-concierge, sets
the VM booting mode active in a while (Currently 60 seconds because low-
end boards will take more than 30 seconds to boot the VM. Double it for
safety.) and when it received 'VmCompleteBootSignal', it set the VM
booting mode inactive.

Slow boards can temporarily set the CPUFreq governor to 'performance'
on this VM boot mode so that it will run in full speed while booting VMs.

This is disabled on x86-64 because if someone accidentally sets the governor
to 'performance' in 1 minutes, the mid to high-end CPUs can be throttled.

BUG=b:264966361
TEST=build for kukui-arc-r and brya

Change-Id: Iad628c0bbc8551ff5a218b0e256aa06c881dceeb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4243469
Tested-by: Masami Hiramatsu <mhiramat@google.com>
Reviewed-by: Vovo Yang <vovoy@chromium.org>
Auto-Submit: Masami Hiramatsu <mhiramat@google.com>
NOKEYCHECK=True
GitOrigin-RevId: 7e76898ad7f4c5770e8e4c4959bf44b3a9e68f7e