featured: Record active trials
Add method that writes active trial files of the form
TrialName,GroupName to /run/featured/active.

There is no logic in this CL to call this method since there is no logic
yet to populate the FeatureOverride proto. This will be implemented in a
future CL for the evaluate_seed binary.

Future work is planned to call this method when feature_library receives
a query for a feature and for Chrome to read the active trial files and
mark any new experiments as active. This is primarily targeted for
early-boot experiments.

More details are available at go/platform-finch-early-proposal. The
relevant section is "Feature_library to chrome communication" on page 8.

BUG=b:281884127
TEST=unit tests

Cq-Depend: chromium:4595237
Change-Id: Ifd2ce368d6ffec85ef8cdef7b6133976554d5aa4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4517959
Tested-by: Kendrake Tsui <kendraketsui@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Luc Nguyen <lucnguyen@google.com>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Commit-Queue: Kendrake Tsui <kendraketsui@google.com>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: 0f5847f3018fb91c9caf375e72d815e3f8f5f72e