system_api: Define cryptohome API for controlling intents
This adds the definition for a future API call that would allow you to
enable or disable the use of particular intents with particular types of
auth factors. We're still quite a way from being able to implement this
(lots of infrastructure bits under the covers) but this gets us ahead in
terms of specifying the API.

This also adds a new field to ListAuthFactors that lays out the options
for each type of factor. This we can implement immediately, since the
longer pole is making things configurable.

BUG=b:281878605
TEST=emerge-$BOARD chromeos-base/system_api

Change-Id: Ie7ef05fc800d918280aba1f8283cc20693b5e86d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4522704
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Denis Kuznetsov <antrim@chromium.org>
Reviewed-by: Behnood Momenzadeh <behnoodm@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: 4e924e6e86a9f9c60b7e17d55ad2879597e77afa