system_api: Drop a bunch of cryptohome boilerplate
In the cryptohome API we have a bunch of boilerplate that we copy into
every Reply message for the error and error_info fields. This isn't
really useful, and in general "copy this comment into almost every other
message" isn't really great for maintenance.

Instead I put a more extensive summary of these fields into the top
level file comment (which is already talking about the Request+Reply
pattern being used) and deleted most of the comments being repeated on
individual fields.

I did leave in place the comments in a few messages where the
information was non-trivial, but if they were just variations on "this
will be set if the operation fails" I dropped them.

BUG=none
TEST=built the lib just to make sure

Change-Id: Id63c7623b27f09f4ac061bb3419ffeedf9339e60
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4534515
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: 44a1464d8724ce051d39f280e9dd018f0008ad10