system_api: attestation: Change attestation protos for new VA flow
Rename KeyProfile and change several attestation protos to be ready for a new VA flow for the Device Trust Connector on unmanaged devices (go/remote-attestation-unmanaged-devices-dd). As part of this, attestation logic will be moved out of the AttestationService. Instead, chromium will be responsible for passing the correct information to the service which is needed to build a valid challenge response for VA.

BUG=b:284099134, b:277709127
TEST=tast run <ip> hwsec.Attestation*

Change-Id: I259c34e8c5ef899f1f237c018d1dc469ee88039a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4584309
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Commit-Queue: Leon Masopust <lmasopust@google.com>
Tested-by: Leon Masopust <lmasopust@google.com>
Reviewed-by: Leo Lai <cylai@google.com>
Reviewed-by: Sébastien Lalancette <seblalancette@chromium.org>
Reviewed-by: Martin Rodriguez <rodmartin@google.com>
NOKEYCHECK=True
GitOrigin-RevId: b66bc363c90cd6bd7ef0da458e3c5d77efe43c40