cryptohome: Remove last of the deprecated authenticated bool
The authenticated boolean is no longer used in the code and is replaced
by authorized for intents. The CL updates the test to reflect that and
also marks it spot in the proto as reserved.

BUG=b:262560281
FIXED=b:262560281
TEST=CQ Passes

Change-Id: I0eb251c12073aa245dc6ccf478201d11a417e496
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4610321
Reviewed-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Auto-Submit: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: e1a37019066fd9b0d8a5ae3a24ad040181e64f84