shill: Add class P2PManager
Add P2PManager class to Shill in the WiFi subdirectory. This is a
skeleton implementation that will be built upon in future development.
Includes stubs for basic functions like Start/Stop, constructors, etc.

Also includes an `allowed` value and set it to false in the
constructor. This is important to include in the skeleton so that we can
leverage the value in all future code changes.

BUG=b:295049819
TEST=FEATURES=test emerge-guybrush shill

Change-Id: I4890cc149ef31a4ba3852f9ac2f5d4dbe6492b94
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4763333
Commit-Queue: Kevin Lund <kglund@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Jacek Siuda <jsiuda@google.com>
NOKEYCHECK=True
GitOrigin-RevId: ada896b900a3019712dd7bc20957610458d583f9