multiple: Refactor fingerprint auth flow
Refactors the fingerprint auth flow. Now most of the steps are moved
from AuthenticateCredential to StartAuthSession. This allows us to
further improve the latency.

Note that this greatly simplifies the code because now we can set the FP
sensor to match mode directly, instead of finger down mode first then
match mode as before.

BUG=b:298346899
TEST=biod unittests
TEST=biod_proxy/cryptohome unittests
TEST=fingerprint login integration tests fail (as expected)

Change-Id: Icfcf3034e4c5dc28429ee86d788d799387454567
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4840804
Reviewed-by: Patryk Duda <patrykd@google.com>
Tested-by: Howard Yang <hcyang@google.com>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: 873326292e831cd1ee9461bc6921baa2c2a7f126