cryptohome: Make ExtendAuthSession do "Extend to"
Here the functionality of ExtendAuthSession is being changed to
extending AuthSession to a particular time as opposed to the original
which was appending the time at the end. The extension duration will be
max of current time left and time requested.

BUG=b:300100967
TEST=FEATURES=test emerge-zork cryptohome

Change-Id: I4308093990a4db1d937db5549c43407c290c86a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4866124
Auto-Submit: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: John Admanski <jadmanski@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: 61bdfc6ddb92cf3552b3e8298a0341ae5cfc0e41