shill: inject the phy index into TetheringManager
When testing the tethering functionality, we need to inject a virtual
WiFi interface as the downstream device, which is not managed by shill.
If the WiFi interface is managed by shill, then the patchpanel will add
some iptables rules to redirect the packets from the interface to ARC.
Then we cannot send the TCP packet from the downstream netns to another
simulated upstream netns.

This CL adds a DBus property to inject the phy index of the downstream
device into TetheringManager. Then the TetheringManager could create the
downstream network on the virtual WiFi interface.

BUG=b:288359516
TEST=FEATURES=test emerge-brya chromeos-base/system_api shill
TEST=pass the tast test locally

Change-Id: Ia2bfad4cbd4a0f77040eb94fc89bcd160cb9f6ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4905053
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: 8e5390838bf84de0ae6cd254194fdcd892c7b3de