shill: cellular: Don't stop round robin on MultipleAccessToPdnConnectionNotAllowed
Continue the connect round robin when ModemManager returns the error
MobileEquipment.MultipleAccessToPdnConnectionNotAllowed.
The error is used to indicate that the network requested the modem to
throttle connection attempts with the current APN.

BUG=b:302722603
TEST=Unit tests pass
Reproduce this on a craask and verify the round robin continues.

Change-Id: I66de58110acc59ce01048d396f8704ded9ae1e4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5254001
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Aleksander Morgado <aleksandermj@google.com>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: b37fb4e6193c6b0f39242a57140901a86c65ebd8