shill: cellular: connect by profile ID if available
Reading profiles out and sending them by their details will
cause issues if e.g. the profile has a specific IP type as
that is dropped when the modem is exposing profiles using
some control planes (like MBIM 1.0). Instead, if the modem gives
us a profile ID, we should use that to identify the profile
directly on the modem, so that we don't lose anything.

BUG=b:310710142
TEST=unit tests, deploy to DUT and verify that connection
  attempts made using profiles will pass the profile ID to
  ModemManager

Change-Id: I49330929dc48a5da94dd948d7747115739776d46
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5067719
Reviewed-by: Aleksander Morgado <aleksandermj@google.com>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: e63c42389847f698de49206bbab65d10ea64a77f