shill: tethering: add Internet validation timer
This patch rehauls the TetheringManager logic checking Internet
connectivity on the upstream Network and replaces it with a strategy
similar to the inactivity timer:
  - The Internet connectivity state of the upstream network is not
    taken into account when starting the Tethering session. It is
    sufficient if the upstream network is connected at the L3 layer.
  - When the Tethering session starts, if Internet access on the
    upstream network has not been validated yet, a 1m timer is fired
  - When Internet access is confirmed on the upstream network (monitored
    via the Network::EventHandler callbacks), the timer is cancelled.
  - If the timer fires, the tethering session is torn down.

The same logic is also used if Internet access is lost in the middle of
a session.

As a consequence, the "upstream_network_without_internet" result for
the Manager DBus method SetTetheringEnabled is now obsolete and replaced
with a new "upstream_no_internet" value. A follow-up patch will start
using that "upstream_no_internet" value for the
kTetheringStatusIdleReasonProperty property once Chrome has been updated
to handle it correctly. For the time being, the "active" reason is used
as a temporary value instead.

BUG=b:271322391
BUG=b:291845893
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: I81c01f6d6e9b124d47d542a5ab016d389b9b25cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5082857
Reviewed-by: Aleksander Morgado <aleksandermj@google.com>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: b962123f1d7574120b7c1286470410e238e006f2